+++ CHANGELOG.md
@@ -7,10 +7,12 @@
 - Replay for crashes (#4171)
 - Redact web view from replay (#4203)
 - Add beforeCaptureViewHierarchy callback (#4210)
+- Rename session replay `errorSampleRate` property to `onErrorSampleRate` (#4218)
 
 ### Fixes
 
 - Skip UI crumbs when target or sender is nil (#4211)
+- Guard FramesTracker start and stop (#4224)
 
 ### Improvements
 
@@ -76,7 +78,7 @@
 ### Fixes
 
 - Fix potential deadlock in app hang detection (#4063)
-- Swizzling of view controllers `loadView` that don`t implement `loadView` (#4071)
+- Swizzling of view controllers `loadView` that don't implement `loadView` (#4071)
 
 ## 8.29.0